requirements.txt
# catalog_dashboard/requirements.txt
streamlit
pandas
requests
PyYAML
gspread
google-auth-oauthlib